# docs/api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: PromptChain API Contract
  version: 1.0.0
  description: |
    PromptChain 统一接口规范（v1）。

    本规范作为前后端联调与后续开发的唯一事实源（SSOT）。
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/trace/{workflow_run_id}:
    get:
      tags: [Trace]
      summary: 获取工作流完整追踪
      operationId: getWorkflowTrace
      parameters:
        - $ref: '#/components/parameters/WorkflowRunId'
      responses:
        '200':
          description: Workflow trace payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowTraceResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/trace/node/{node_run_id}:
    get:
      tags: [Trace]
      summary: 获取节点执行详情
      operationId: getNodeDetail
      parameters:
        - $ref: '#/components/parameters/NodeRunId'
      responses:
        '200':
          description: Node detail payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/artifact/{artifact_id}:
    get:
      tags: [Artifact]
      summary: 获取单个 Artifact
      operationId: getArtifact
      parameters:
        - $ref: '#/components/parameters/ArtifactId'
      responses:
        '200':
          description: Artifact detail
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
  /api/artifact/{artifact_id}/history:
    get:
      tags: [Artifact]
      summary: 获取 Artifact 版本历史
      operationId: getArtifactHistory
      parameters:
        - $ref: '#/components/parameters/ArtifactId'
      responses:
        '200':
          description: Artifact history list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        '500':
          $ref: '#/components/responses/InternalError'
  /api/workflows:
    get:
      tags: [WorkflowDefinition]
      summary: 获取当前工作空间工作流定义列表
      operationId: listWorkflowDefinitions
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Result envelope with workflow list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [WorkflowDefinition]
      summary: 创建工作流定义
      operationId: createWorkflowDefinition
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowDefinitionCreate'
      responses:
        '200':
          description: Result envelope with created workflow definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/workflows/{workflow_id}/versions:
    get:
      tags: [WorkflowVersion]
      summary: 获取工作流版本历史
      operationId: listWorkflowVersions
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: Result envelope with version list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/workflows/{workflow_id}/versions/{version_id}:
    get:
      tags: [WorkflowVersion]
      summary: 获取指定版本定义
      operationId: getWorkflowVersion
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: Result envelope with version payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/workflows/{workflow_id}/versions/compare:
    get:
      tags: [WorkflowVersion]
      summary: 对比两个版本差异
      operationId: compareWorkflowVersions
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
        - name: version_a
          in: query
          required: true
          schema:
            type: integer
        - name: version_b
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Result envelope with diff payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/workflows/{workflow_id}/versions/{version_id}/restore:
    post:
      tags: [WorkflowVersion]
      summary: 回滚到指定版本
      operationId: restoreWorkflowVersion
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: Result envelope with restore payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  parameters:
    WorkflowRunId:
      name: workflow_run_id
      in: path
      required: true
      schema:
        type: string
      description: 工作流运行 ID
    WorkflowId:
      name: workflow_id
      in: path
      required: true
      schema:
        type: string
      description: 工作流定义 ID
    VersionId:
      name: version_id
      in: path
      required: true
      schema:
        type: string
      description: 版本 ID
    NodeRunId:
      name: node_run_id
      in: path
      required: true
      schema:
        type: string
      description: 节点运行 ID
    ArtifactId:
      name: artifact_id
      in: path
      required: true
      schema:
        type: string
      description: Artifact ID
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
        code:
          type: string
          nullable: true
...
    ResultResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
...
    WorkflowTraceResponse:
      type: object
      required: [workflow, nodes, artifacts, timeline]
      properties:
        workflow:
          type: object
          additionalProperties: true
        nodes:
          type: array
          items:
            type: object
            additionalProperties: true
        artifacts:
          type: object
          additionalProperties:
            type: object
            additionalProperties: true
        timeline:
          type: array
          items:
            type: object
            additionalProperties: true
    NodeDetailResponse:
      type: object
      required: [node, input_artifacts, output_artifacts]
      properties:
        node:
          type: object
          additionalProperties: true
        input_artifacts:
          type: array
          items:
            type: object
            additionalProperties: true
        output_artifacts:
          type: array
          items:
            type: object
            additionalProperties: true
    NodePosition:
      type: object
      required: [x, y]
      properties:
        x:
          type: number
        y:
          type: number
    NodeData:
      type: object
      required: [label]
      properties:
        label:
          type: string
        config:
          type: object
          additionalProperties: true
    WorkflowNode:
      type: object
      required: [id, type, position, data]
      properties:
        id:
          type: string
        type:
          type: string
        position:
          $ref: '#/components/schemas/NodePosition'
        data:
          $ref: '#/components/schemas/NodeData'
    EdgeData:
      type: object
      properties:
        condition:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
    WorkflowEdge:
      type: object
      required: [id, source, target]
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        type:
          type: string
          default: default
        data:
          $ref: '#/components/schemas/EdgeData'
    WorkflowDefinitionCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          default: ""
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowNode'
          default: []
        edges:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowEdge'
          default: []
...